requirements.txt
selenium
pytest
allure-pytest
PyYAML